requirements.txt
flask
bcrypt
Flask-Session
pillow